Move marker tests into marker package
