+++ src/test/kotlin/io/github/orangain/jsonmatch/marker/ArrayBracketMarkerWithNestedChildrenTest.kt
@@ -1,5 +1,6 @@
-package io.github.orangain.jsonmatch
+package io.github.orangain.jsonmatch.marker
 
+import io.github.orangain.jsonmatch.JsonStringAssert
 import org.assertj.core.api.Assertions
 import org.junit.Test
 
@@ -9,7 +10,8 @@ class ArrayBracketMarkerWithNestedChildrenTest {
     @Test
     fun matchNestedNumberArray() {
         // language=JSON
-        JsonStringAssert.assertThat("""{ "a": [[1, 2, 3], [4, 5, 6]] }""").jsonMatches(patternJson)
+        JsonStringAssert.assertThat("""{ "a": [[1, 2, 3], [4, 5, 6]] }""")
+            .jsonMatches(patternJson)
     }
 
     @Test
@@ -37,7 +39,8 @@ class ArrayBracketMarkerWithNestedChildrenTest {
     fun doesNotMatchNestedArrayContainsString() {
         Assertions.assertThatThrownBy {
             // language=JSON
-            JsonStringAssert.assertThat("""{ "a": [["1", "2", "3"]] }""").jsonMatches(patternJson)
+            JsonStringAssert.assertThat("""{ "a": [["1", "2", "3"]] }""")
+                .jsonMatches(patternJson)
         }.isInstanceOf(AssertionError::class.java)
             .hasMessageContaining("path: $[0], actual: '1', expected: '#number', reason: not a number")
     }
