+++ src/test/kotlin/io/github/orangain/jsonmatch/marker/ArrayMarkerTest.kt
@@ -1,10 +1,11 @@
-package io.github.orangain.jsonmatch
+package io.github.orangain.jsonmatch.marker
 
+import io.github.orangain.jsonmatch.JsonStringAssert
 import org.assertj.core.api.Assertions
 import org.junit.Test
 
-class ObjectMarkerTest {
-    private val patternJson = """{ "a": "#object" }"""
+class ArrayMarkerTest {
+    private val patternJson = """{ "a": "#array" }"""
 
     @Test
     fun doesNotMatchMissingField() {
@@ -12,7 +13,7 @@ class ObjectMarkerTest {
             // language=JSON
             JsonStringAssert.assertThat("{}").jsonMatches(patternJson)
         }.isInstanceOf(AssertionError::class.java)
-            .hasMessageContaining("path: \$, actual: {}, expected: {a=#object}, reason: all key-values did not match, expected has un-matched keys: [a]")
+            .hasMessageContaining("path: \$, actual: {}, expected: {a=#array}, reason: all key-values did not match, expected has un-matched keys: [a]")
     }
 
     @Test
@@ -21,7 +22,7 @@ class ObjectMarkerTest {
             // language=JSON
             JsonStringAssert.assertThat("""{ "a": null }""").jsonMatches(patternJson)
         }.isInstanceOf(AssertionError::class.java)
-            .hasMessageContaining("path: $.a, actual: null, expected: '#object', reason: not a json object")
+            .hasMessageContaining("path: $.a, actual: null, expected: '#array', reason: not an array or list")
     }
 
     @Test
@@ -30,7 +31,7 @@ class ObjectMarkerTest {
             // language=JSON
             JsonStringAssert.assertThat("""{ "a": true }""").jsonMatches(patternJson)
         }.isInstanceOf(AssertionError::class.java)
-            .hasMessageContaining("path: $.a, actual: true, expected: '#object', reason: not a json object")
+            .hasMessageContaining("path: $.a, actual: true, expected: '#array', reason: not an array or list")
     }
 
     @Test
@@ -39,7 +40,7 @@ class ObjectMarkerTest {
             // language=JSON
             JsonStringAssert.assertThat("""{ "a": 1 }""").jsonMatches(patternJson)
         }.isInstanceOf(AssertionError::class.java)
-            .hasMessageContaining("path: $.a, actual: 1, expected: '#object', reason: not a json object")
+            .hasMessageContaining("path: $.a, actual: 1, expected: '#array', reason: not an array or list")
     }
 
     @Test
@@ -48,27 +49,27 @@ class ObjectMarkerTest {
             // language=JSON
             JsonStringAssert.assertThat("""{ "a": "true" }""").jsonMatches(patternJson)
         }.isInstanceOf(AssertionError::class.java)
-            .hasMessageContaining("path: $.a, actual: 'true', expected: '#object', reason: not a json object")
+            .hasMessageContaining("path: $.a, actual: 'true', expected: '#array', reason: not an array or list")
     }
 
     @Test
-    fun doesNotMatchArray() {
+    fun matchArray() {
+        // language=JSON
+        JsonStringAssert.assertThat("""{ "a": [1, 2, 3] }""").jsonMatches(patternJson)
+    }
+
+    @Test
+    fun matchEmptyArray() {
+        // language=JSON
+        JsonStringAssert.assertThat("""{ "a": [] }""").jsonMatches(patternJson)
+    }
+
+    @Test
+    fun doesNotMatchObject() {
         Assertions.assertThatThrownBy {
             // language=JSON
-            JsonStringAssert.assertThat("""{ "a": [] }""").jsonMatches(patternJson)
+            JsonStringAssert.assertThat("""{ "a": {} }""").jsonMatches(patternJson)
         }.isInstanceOf(AssertionError::class.java)
-            .hasMessageContaining("path: $.a, actual: [], expected: '#object', reason: not a json object")
-    }
-
-    @Test
-    fun matchObject() {
-        // language=JSON
-        JsonStringAssert.assertThat("""{ "a": { "b": 1 } }""").jsonMatches(patternJson)
-    }
-
-    @Test
-    fun matchEmptyObject() {
-        // language=JSON
-        JsonStringAssert.assertThat("""{ "a": {} }""").jsonMatches(patternJson)
+            .hasMessageContaining("path: $.a, actual: {}, expected: '#array', reason: not an array or list")
     }
 }
