+++ src/test/kotlin/io/github/orangain/jsonmatch/marker/IgnoreMarkerTest.kt
@@ -1,5 +1,6 @@
-package io.github.orangain.jsonmatch
+package io.github.orangain.jsonmatch.marker
 
+import io.github.orangain.jsonmatch.JsonStringAssert
 import org.junit.Test
 
 class IgnoreMarkerTest {
