+++ src/test/kotlin/io/github/orangain/jsonmatch/marker/NumberMarkerTest.kt
@@ -1,5 +1,6 @@
-package io.github.orangain.jsonmatch
+package io.github.orangain.jsonmatch.marker
 
+import io.github.orangain.jsonmatch.JsonStringAssert
 import org.assertj.core.api.Assertions
 import org.junit.Test
